requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0

# LangChain for AI model integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0

# Environment and configuration
python-dotenv>=1.0.0

# Additional utilities
requests